test(pnpr): reuse production lock mapping
